Fix PRAGMA key or salt has incorrect value when generating groups
