+++ xmtp_debug/src/app/generate/messages.rs
@@ -1,3 +1,4 @@
+use crate::app::identity_lock::get_identity_lock;
 use crate::{
     app::{
         self,
@@ -8,27 +9,10 @@ use crate::{
 use color_eyre::eyre::{self, Result, eyre};
 use indicatif::{ProgressBar, ProgressStyle};
 use rand::{Rng, SeedableRng, rngs::SmallRng, seq::SliceRandom};
-use std::collections::HashMap;
-use std::sync::{Arc, Mutex, OnceLock};
-use tokio::sync::Mutex as TokioMutex;
+use std::sync::Arc;
 use xmtp_mls::groups::summary::SyncSummary;
 
 mod content_type;
-
-type IdentityLockMap = Arc<Mutex<HashMap<[u8; 32], Arc<TokioMutex<()>>>>>;
-
-static IDENTITY_LOCKS: OnceLock<IdentityLockMap> = OnceLock::new();
-
-fn get_identity_lock(inbox_id: &[u8; 32]) -> Result<Arc<TokioMutex<()>>, eyre::Error> {
-    let locks = IDENTITY_LOCKS.get_or_init(|| Arc::new(Mutex::new(HashMap::new())));
-    let mut map = locks
-        .lock()
-        .map_err(|e| eyre!("Failed to lock IDENTITY_LOCKS: {}", e))?;
-    Ok(map
-        .entry(*inbox_id)
-        .or_insert_with(|| Arc::new(TokioMutex::new(())))
-        .clone())
-}
 
 #[derive(thiserror::Error, Debug)]
 enum MessageSendError {
